+++ src/main/java/org/patternfly/component/alert/Alert.java
@@ -61,7 +61,6 @@ import static org.patternfly.handler.CloseHandler.fireEvent;
 import static org.patternfly.handler.CloseHandler.shouldClose;
 import static org.patternfly.style.Classes.alert;
 import static org.patternfly.style.Classes.component;
-import static org.patternfly.style.Classes.expandable;
 import static org.patternfly.style.Classes.icon;
 import static org.patternfly.style.Classes.modifier;
 import static org.patternfly.style.Classes.screenReader;
@@ -96,6 +95,7 @@ public class Alert extends BaseComponent<HTMLDivElement, Alert> implements
     static final int MIN_TIMEOUT = 1_000; // ms
 
     int timeout;
+    boolean expandable;
     Button closeButton;
     CloseHandler<Alert> closeHandler;
     private final Severity severity;
@@ -114,6 +114,7 @@ public class Alert extends BaseComponent<HTMLDivElement, Alert> implements
         this.severity = severity;
         this.title = title;
         this.timeout = NO_TIMEOUT;
+        this.expandable = false;
         this.timeoutHandle = 0;
 
         add(iconContainer = div().css(component(alert, icon))
@@ -162,7 +163,7 @@ public class Alert extends BaseComponent<HTMLDivElement, Alert> implements
     // override to assure internal wiring
     public Alert add(AlertDescription description) {
         this.description = description;
-        this.description.element().hidden = element().classList.contains(modifier(expandable)) && !expanded();
+        this.description.element().hidden = element().classList.contains(modifier(Classes.expandable)) && !expanded();
         add(description.element());
         return this;
     }
@@ -200,7 +201,7 @@ public class Alert extends BaseComponent<HTMLDivElement, Alert> implements
     }
 
     public Alert expandable(ToggleHandler<Alert> toggleHandler) {
-        css(modifier(expandable));
+        css(modifier(Classes.expandable));
         insertFirst(element(), div().css(component(alert, toggle))
                 .add(toggleButton = button().plain()
                         .on(click, e -> toggle())
@@ -209,6 +210,7 @@ public class Alert extends BaseComponent<HTMLDivElement, Alert> implements
                         .add(span().css(component(alert, toggle, icon))
                                 .add(inlineIcon(angleRight))))
                 .element());
+        this.expandable = true;
         this.toggleHandler = toggleHandler;
         return this;
     }
@@ -299,6 +301,11 @@ public class Alert extends BaseComponent<HTMLDivElement, Alert> implements
 
     @Override
     public void collapse(boolean fireEvent) {
+        if (!expandable) {
+            Logger.unsupported(ComponentType.Alert, element(), "Alert is not expandable.\n" +
+                    "Please use Alert.expandable() to make this an expandable alert.");
+            return;
+        }
         Expandable.collapse(element(), toggleButton.element(), description.element());
         if (fireEvent && toggleHandler != null) {
             toggleHandler.onToggle(new Event(""), this, false);
@@ -307,6 +314,11 @@ public class Alert extends BaseComponent<HTMLDivElement, Alert> implements
 
     @Override
     public void expand(boolean fireEvent) {
+        if (!expandable) {
+            Logger.unsupported(ComponentType.Alert, element(), "Alert is not expandable.\n" +
+                    "Please use Alert.expandable() to make this an expandable alert.");
+            return;
+        }
         Expandable.expand(element(), toggleButton.element(), description.element());
         if (fireEvent && toggleHandler != null) {
             toggleHandler.onToggle(new Event(""), this, true);