+++ src/main/java/org/patternfly/layout/grid/GridItem.java
@@ -16,12 +16,16 @@
 package org.patternfly.layout.grid;
 
 import org.patternfly.core.Logger;
+import org.patternfly.core.Tuple;
+import org.patternfly.core.Tuples;
 import org.patternfly.layout.BaseLayout;
+import org.patternfly.style.Breakpoint;
 import org.patternfly.style.Modifiers.Fill;
 
 import elemental2.dom.HTMLDivElement;
 
 import static org.jboss.elemento.Elements.div;
+import static org.patternfly.layout.grid.Grid.internalOrder;
 import static org.patternfly.style.Classes.grid;
 import static org.patternfly.style.Classes.item;
 import static org.patternfly.style.Classes.layout;
@@ -43,16 +47,80 @@ public class GridItem extends BaseLayout<HTMLDivElement, GridItem> implements Fi
 
     // ------------------------------------------------------ builder
 
-    public GridItem colSpan(int cols) {
-        if (verifyRange("colSpan", cols)) {
-            css(modifier(cols + "-col"));
+    /**
+     * The number of columns the grid item spans. Value should be a number 1-12.
+     */
+    public GridItem span(int columns) {
+        if (verifyRange("span", columns)) {
+            css(modifier(columns + "-col"));
         }
         return this;
     }
 
+    /**
+     * The number of columns the grid item spans on a specific breakpoint. Value should be a number 1-12.
+     */
+    public GridItem span(Tuples<Breakpoint, Integer> columns) {
+        if (columns != null) {
+            for (Tuple<Breakpoint, Integer> column : columns) {
+                internalSpan(column);
+            }
+        }
+        return this;
+    }
+
+    /**
+     * The number of rows the grid item spans. Value should be a number 1-12.
+     */
     public GridItem rowSpan(int rows) {
         if (verifyRange("rowSpan", rows)) {
             css(modifier(rows + "-row"));
+        }
+        return this;
+    }
+
+    /**
+     * The number of rows the grid item spans on a specific breakpoint. Value should be a number 1-12
+     */
+    public GridItem rowSpan(Tuples<Breakpoint, Integer> rows) {
+        if (rows != null) {
+            for (Tuple<Breakpoint, Integer> row : rows) {
+                internalRowSpan(row);
+            }
+        }
+        return this;
+    }
+
+    /**
+     * The number of columns a grid item is offset.
+     */
+    public GridItem offset(int columns) {
+        if (verifyRange("offset", columns)) {
+            css(modifier("offset-" + columns + "-col"));
+        }
+        return this;
+    }
+
+    /**
+     * The number of columns the grid item is offset on a specific breakpoint. Value should be a number 1-12
+     */
+    public GridItem offset(Tuples<Breakpoint, Integer> columns) {
+        if (columns != null) {
+            for (Tuple<Breakpoint, Integer> column : columns) {
+                internalOffset(column);
+            }
+        }
+        return this;
+    }
+
+    /**
+     * Modifies the flex layout element order property.
+     */
+    public GridItem order(Tuples<Breakpoint, String> order) {
+        if (order != null) {
+            for (Tuple<Breakpoint, String> o : order) {
+                internalOrder(element(), o);
+            }
         }
         return this;
     }
@@ -63,6 +131,30 @@ public class GridItem extends BaseLayout<HTMLDivElement, GridItem> implements Fi
     }
 
     // ------------------------------------------------------ internal
+
+    private void internalSpan(Tuple<Breakpoint, Integer> tuple) {
+        if (tuple.key != Breakpoint.default_) {
+            if (verifyRange("span", tuple.value)) {
+                css(modifier(tuple.value + "-col-on-" + tuple.key.value));
+            }
+        }
+    }
+
+    private void internalRowSpan(Tuple<Breakpoint, Integer> tuple) {
+        if (tuple.key != Breakpoint.default_) {
+            if (verifyRange("rowSpan", tuple.value)) {
+                css(modifier(tuple.value + "-row-on-" + tuple.key.value));
+            }
+        }
+    }
+
+    private void internalOffset(Tuple<Breakpoint, Integer> tuple) {
+        if (tuple.key != Breakpoint.default_) {
+            if (verifyRange("offset", tuple.value)) {
+                css(modifier("offset-" + tuple.value + "-col-on-" + tuple.key.value));
+            }
+        }
+    }
 
     private boolean verifyRange(String property, int value) {
         if (value < 1 || value > 12) {